--- crates/kubelet/src/example-bootstrap-kubelet.conf
@@ -1,18 +1,0 @@
-apiVersion: v1
-clusters:
-- cluster:
-    certificate-authority-data: LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUN5RENDQWJDZ0F3SUJBZ0lCQURBTkJna3Foa2lHOXcwQkFRc0ZBREFWTVJNd0VRWURWUVFERXdwcmRXSmwKY201bGRHVnpNQjRYRFRJd01EVXlNakV4TURRek5sb1hEVE13TURVeU1ERXhNRFF6Tmxvd0ZURVRNQkVHQTFVRQpBeE1LYTNWaVpYSnVaWFJsY3pDQ0FTSXdEUVlKS29aSWh2Y05BUUVCQlFBRGdnRVBBRENDQVFvQ2dnRUJBTXhnCi81dW5RUC9WbVhZUkV4Um5KRkhyRmpZU3VaU1dtL1o4Vlp4RlJjemwxbmVwYUI4c0xVR25KNUNTVGVjUVY5d0kKZmc3NmoweXdta2toQktNdGllVFVTSHE4TTJxbW83aG9RSlhaQytZQ0x3WXRiR0U3eW9CN0o3cUg0WVh4Qk9LSQo5dGh0Q3h2Y096YjRwbGNqdSs2NVhaaW5DaGVsNDh4WHMzeDk0R2dGdTNtb2pLbTdOVVRwL0NNclV5WEVwR1hHClFBYTJOV05STjgyLys1L2wzOXhCVTYzOXA5RjhpaEkvK3pLTEhYZlNIenBDNkV0OFB0dlhmOGxJUTEvRDJtYjUKU04zNlV0NmtFYmt6b2NVMVUzNVpZdHhKaStLSkZjbDBkSUFpb0JrK2dyb09mcHhnV0U0RlYrU1A0Z0NJaHNaWgo1TkhDMXBuVkoxZ2Ziby8rRmdrQ0F3RUFBYU1qTUNFd0RnWURWUjBQQVFIL0JBUURBZ0trTUE4R0ExVWRFd0VCCi93UUZNQU1CQWY4d0RRWUpLb1pJaHZjTkFRRUxCUUFEZ2dFQkFEWWhodGtoQ29JQUMyVUZld1FEZDI2ZW9qYUUKSVVUQUpTdEFWUVNpOFRHbi9aMlV2Z1lQOE52MFl4cVVod2tXajhRYytaK2srTlRYcVhMR1FEeDhYK2h1UWExSgowM3VoRlNNMkhJeDRQVy80K1VTbml1Sm1aY3BWdGlkNDBDYlRCbzdUdWJudGVSZkVtdW1WOHFHbkJFTEljM09XCktRb3dEMWgycVd5bDk2YmhBb1BwN05pNnNHYmhhNjdlNldObEcxdDJmSHJUWVZ4Nm03cWd4OUw3dzVpWlZ0cE4KRXpCR1VXczJUM2NyRk9WTzF2UUdkTU04aXBKY1VESENhMnJHU0NMQU5zeHlaOFZ6RSs5ZHZWYnVJMWlMMS9SUwp3Y0JMbUtXTW5oVWZvU0VPYlBlbTVZU09UR2NmZ0F6czU5REpjWENYMHJJNEtWazNvM1poaW0yeTMrbz0KLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQo=
-    server: http://127.0.0.1:6443
-  name: kubernetes
-contexts:
-- context:
-    cluster: kubernetes
-    user: tls-bootstrap-token-user
-  name: tls-bootstrap-token-user@kubernetes
-current-context: tls-bootstrap-token-user@kubernetes
-kind: Config
-preferences: {}
-users:
-- name: tls-bootstrap-token-user
-  user:
-    token: zrgtfy.09fzbp9q1ae88x26
